Add files via upload
Slight changes to the GUI: replaced DEL button with CE (clear-entry) and C (clear-all) buttons.

Under the hood, added functionality behind CE and C buttons (in form of onClick attribute in the html tags).
Added universal function for typing in numbers and a comma (numberpad).
Added functionality for operation buttons as Event Listeners (not sure if this is the most efficient way, or whether it's best to do it with html onclick attribute).

Things to resolve: when operator is in the entry field, inputting numbers does not automatically move the operator to the middle window (when it should); when there are two values and operator in the middle, currently only pressing equals solves the equation. Ideally, all operators should be able to do it (perhaps insert the equals function with a conditional inside the operations functions?).